pyproject.toml
[project]
name = "mit"
version = "0.1.0"
description = "Multi Intelligence Twin - A multi-agent framework for expertise"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mit"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
